pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord>=2.3.2",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.39.0",
]
